+++ packages/integration/src/gitlab/GitLabIntegration.test.ts
@@ -20,21 +20,16 @@ import { GitLabIntegration } from './GitLabIntegration';
 describe('GitLabIntegration', () => {
   it('has a working factory', () => {
     const integrations = GitLabIntegration.factory({
-      config: ConfigReader.fromConfigs([
-        {
-          context: '',
-          data: {
-            integrations: {
-              gitlab: [
-                {
-                  host: 'h.com',
-                  token: 't',
-                },
-              ],
+      config: new ConfigReader({
+        integrations: {
+          gitlab: [
+            {
+              host: 'h.com',
+              token: 't',
             },
-          },
+          ],
         },
-      ]),
+      }),
     });
     expect(integrations.length).toBe(2); // including default
     expect(integrations[0].predicate(new URL('https://h.com/a'))).toBe(true);