+++ plugins/scaffolder-backend/src/scaffolder/stages/prepare/gitlab.test.ts
@@ -42,7 +42,7 @@ const mockEntityWithProtocol = (protocol: string): TemplateEntityV1alpha1 => ({
     name: 'graphql-starter',
     title: 'GraphQL Service',
     description:
-      'A GraphQL starter template for backstage to get you up and running\nthe best pracices with GraphQL\n',
+      'A GraphQL starter template for backstage to get you up and running\nthe best practices with GraphQL\n',
     uid: '9cf16bad-16e0-4213-b314-c4eec773c50b',
     etag: 'ZTkxMjUxMjUtYWY3Yi00MjU2LWFkYWMtZTZjNjU5ZjJhOWM2',
     generation: 1,
@@ -78,7 +78,7 @@ describe('GitLabPreparer', () => {
 
   ['gitlab', 'gitlab/api'].forEach(protocol => {
     it(`calls the clone command with the correct arguments for a repository using the ${protocol} protocol`, async () => {
-      const preparer = new GitlabPreparer(ConfigReader.fromConfigs([]));
+      const preparer = new GitlabPreparer(new ConfigReader({}));
       mockEntity = mockEntityWithProtocol(protocol);
       await preparer.prepare(mockEntity, { logger: getVoidLogger() });
       expect(mocks.Clone.clone).toHaveBeenNthCalledWith(
@@ -89,22 +89,42 @@ describe('GitLabPreparer', () => {
       );
     });
 
-    it(`calls the clone command with the correct arguments if an access token is provided for a repository using the ${protocol} protocol`, async () => {
+    it(`calls the clone command with the correct arguments if an access token is provided in integrations for a repository using the ${protocol} protocol`, async () => {
       const preparer = new GitlabPreparer(
-        ConfigReader.fromConfigs([
-          {
-            context: '',
-            data: {
-              catalog: {
-                processors: {
-                  gitlabApi: {
-                    privateToken: 'fake-token',
-                  },
-                },
+        new ConfigReader({
+          integrations: {
+            gitlab: [
+              {
+                host: 'gitlab.com',
+                token: 'fake-token',
               },
+            ],
+          },
+        }),
+      );
+      mockEntity = mockEntityWithProtocol(protocol);
+      await preparer.prepare(mockEntity, { logger: getVoidLogger() });
+      expect(mocks.Clone.clone).toHaveBeenNthCalledWith(
+        1,
+        'https://gitlab.com/benjdlambert/backstage-graphql-template',
+        expect.any(String),
+        {
+          fetchOpts: {
+            callbacks: {
+              credentials: expect.anything(),
             },
           },
-        ]),
+        },
+      );
+    });
+
+    it(`calls the clone command with the correct arguments if an access token is provided in scaffolder for a repository using the ${protocol} protocol`, async () => {
+      const preparer = new GitlabPreparer(
+        new ConfigReader({
+          scaffolder: {
+            gitlab: { api: { token: 'fake-token' } },
+          },
+        }),
       );
       mockEntity = mockEntityWithProtocol(protocol);
       await preparer.prepare(mockEntity, { logger: getVoidLogger() });
@@ -123,7 +143,7 @@ describe('GitLabPreparer', () => {
     });
 
     it(`calls the clone command with the correct arguments for a repository when no path is provided using the ${protocol} protocol`, async () => {
-      const preparer = new GitlabPreparer(ConfigReader.fromConfigs([]));
+      const preparer = new GitlabPreparer(new ConfigReader({}));
       mockEntity = mockEntityWithProtocol(protocol);
       delete mockEntity.spec.path;
       await preparer.prepare(mockEntity, { logger: getVoidLogger() });
@@ -136,7 +156,7 @@ describe('GitLabPreparer', () => {
     });
 
     it(`return the temp directory with the path to the folder if it is specified using the ${protocol} protocol`, async () => {
-      const preparer = new GitlabPreparer(ConfigReader.fromConfigs([]));
+      const preparer = new GitlabPreparer(new ConfigReader({}));
       mockEntity = mockEntityWithProtocol(protocol);
       mockEntity.spec.path = './template/test/1/2/3';
       const response = await preparer.prepare(mockEntity, {
@@ -149,7 +169,7 @@ describe('GitLabPreparer', () => {
     });
 
     it('return the working directory with the path to the folder if it is specified', async () => {
-      const preparer = new GitlabPreparer(ConfigReader.fromConfigs([]));
+      const preparer = new GitlabPreparer(new ConfigReader({}));
       mockEntity.spec.path = './template/test/1/2/3';
       const response = await preparer.prepare(mockEntity, {
         logger: getVoidLogger(),