+++ plugins/catalog-import/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @backstage/plugin-catalog-import
 
+## 0.3.2
+
+### Patch Changes
+
+- f3e75508d: Add description to Pull Request when registering a new component
+- Updated dependencies [c6eeefa35]
+- Updated dependencies [fb386b760]
+- Updated dependencies [c911061b7]
+- Updated dependencies [7c3ffc0cd]
+- Updated dependencies [e7496dc3e]
+- Updated dependencies [8ef71ed32]
+- Updated dependencies [1d1c2860f]
+- Updated dependencies [0e6298f7e]
+- Updated dependencies [8dd0a906d]
+- Updated dependencies [4eafdec4a]
+- Updated dependencies [6b37c95bf]
+- Updated dependencies [8c31c681c]
+- Updated dependencies [7b98e7fee]
+- Updated dependencies [178e09323]
+- Updated dependencies [ac3560b42]
+- Updated dependencies [0097057ed]
+  - @backstage/plugin-catalog-backend@0.5.0
+  - @backstage/catalog-model@0.6.0
+  - @backstage/core@0.4.1
+  - @backstage/integration@0.1.4
+  - @backstage/plugin-catalog@0.2.7
+
 ## 0.3.1
 
 ### Patch Changes